refactor(ci): Remove unused path filters from workflow
The `paths` filter in the GitHub Actions workflow for Python tests is no longer necessary. Removing it ensures the workflow runs on all pushes to `main` and `dev` branches, regardless of file changes.

The `useAuth` hook import path has been corrected in the frontend components to reflect its new location in `src/hooks/`.
