+++ .github/workflows/python_tests.yml
@@ -3,24 +3,8 @@ name: Python Tests
 on:
   push:
     branches: [ "main", "dev" ]
-    paths:
-      - 'bot.py'
-      - 'cogs/**'
-      - 'tests/**'
-      - 'dashboard/backend/**'
-      - 'database/**'
-      - 'pyrightconfig.json'
-      - 'requirements.txt'
   pull_request:
     branches: [ "main", "dev" ]
-    paths:
-      - 'bot.py'
-      - 'cogs/**'
-      - 'tests/**'
-      - 'dashboard/backend/**'
-      - 'database/**'
-      - 'pyrightconfig.json'
-      - 'requirements.txt'
 
 jobs:
   bot_tests:
